Fixed warning on windows.
git-svn-id: https://svn.opendtect.org/od/trunk@27050 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Algo/spectrogram.cc
@@ -58,7 +58,7 @@ bool Spectrogram::transform( const ArrayND< float > &in, ArrayND< float > &out )
 
     float_complex* tindata = tempin_->getData();
     const float* indata = in.getData();
-    unsigned long size = info.getTotalSz();
+    unsigned long size = (unsigned int) info.getTotalSz();
 
     for ( unsigned int idx=0; idx<size; idx++ )
 	tindata[idx] = indata[idx];
@@ -86,7 +86,7 @@ bool Spectrogram::transform( const ArrayND< float_complex > &in,
     fft_->setOutput( out.getData() );
     if ( !fft_->run(true) ) return false;
 
-    unsigned long size = in.info().getTotalSz();
+    unsigned long size = (unsigned int) in.info().getTotalSz();
     float_complex* data = out.getData();
 
     for ( unsigned int idx=0; idx<size; idx++ )
